+++ TameableCreatures/Properties/AssemblyInfo.cs
@@ -2,11 +2,12 @@
 
 [assembly: AssemblyCompany("TameableCreatures")]
 [assembly: AssemblyConfiguration("Release")]
-[assembly: AssemblyFileVersion("0.13.0.0")]
-[assembly: AssemblyInformationalVersion("0.13.0")]
+[assembly: AssemblyFileVersion("0.14.0.0")]
+[assembly: AssemblyInformationalVersion("0.14.0")]
 [assembly: AssemblyProduct("TameableCreatures")]
 [assembly: AssemblyTitle("TameableCreatures")]
-[assembly: AssemblyVersion("0.13.0.0")]
+[assembly: AssemblyVersion("0.14.0.0")]
+